+++ backend/controllers/request.go
@@ -21,8 +21,9 @@ func RequestToServiceAnalytics(ctx *gin.Context, service models.Service, inputDa
 	var err error
 
 	log.WithFields(log.Fields{
-		"data_service": service,
-		"data_input":   inputData,
+		"data_service":      service,
+		"session_id":        inputData.SessionID,
+		"additional_params": inputData.Data.AdditionalParams,
 	}).Info("[CONTROLLER: RequestToServiceAnalytics] request to analytics service start...")
 
 	if service.Slug == "face-match-enrollment" {
@@ -68,8 +69,9 @@ func RequestToServiceInnovation(ctx *gin.Context, service models.Service, inputD
 	var serviceData models.ServiceRequestResultData
 
 	log.WithFields(log.Fields{
-		"data_service": service,
-		"data_input":   inputData,
+		"data_service":      service,
+		"session_id":        inputData.SessionID,
+		"additional_params": inputData.Data.AdditionalParams,
 	}).Info("[CONTROLLER: RequestToServiceInnovation] request to innovation service start...")
 
 	requestData := inputData.Data